expose attributes on readdir thrift endpoint
Summary:
We want to be able to expose attribute information for all the entries in a
directory with the readdir endpoint. This will allow Buck2 to access attributes
it needs while listing directories (filetype) and experiment with fetching all
the data it ever needs at once.

Now we have all the data we need exposed, we just need to transform it into
thrift types and send it off.

Reviewed By: chadaustin, xavierd

Differential Revision: D37487921

fbshipit-source-id: b265665ae5b771a7a89906bb94783f1be9f85923