get attributes for InodeOr
Summary:
We want to be able to expose attribute information for all the entries in a
directory with the readdir endpoint. This will allow Buck2 to access attributes
it needs while listing directories (filetype) and experiment with fetching all
the data it ever needs at once.

Now we have InodeOrTreeOrEntry for each child we can retrive attribute
information from here for each child.

note: this two phase collection of data looks safe to
renames/deletions/additions. The InodeOrTreeOrEntry have sufficient information
to lookup attributes directly. So we should not end up in a situation where
we get an entry, but then can not lookup attributes for it. The children we list
will be from a consitent state for the directory, but the data may be a bit
stale if there are concurrent modifications to the inodes (again renames,
removals, additons).

Reviewed By: xavierd

Differential Revision: D37444846

fbshipit-source-id: 3c7f33bbec60fb83e4a46403dce715a188ab47f3