InodeOrTreeOrEntry: getInodeOrTreeOrEntry for each entry
Summary:
We want to be able to expose attribute information for all the entries in a
directory with the readdir endpoint. This will allow Buck2 to access attributes
it needs while listing directories (filetype) and experiment with fetching all
the data it ever needs at once.

The first step towards this is fetching all the child InodeOrTreeOrEntrys for
a directory. TreeInode now exposes this information, so we now use this to
expose all child InodeOrTreeOrEntrys for a directory one level up.

Reviewed By: chadaustin, xavierd

Differential Revision: D37259420

fbshipit-source-id: ca85c8e57153cdf94e8f924040aab7df4ae9df9f