Subscribe ActivityBuffer to inodeTraceBus
Summary:
In recent diffs we set up an InodeTraceBus in each EdenMount to receive published Inode materialization events and support a live eden trace inode command (D37528930 (https://github.com/facebookexperimental/eden/commit/72c177c337ee310f0e6926d9b651b099a717bfbb)) as well as next refactor the ActivityBuffer data structure in EdenFS to utilize the InodeTraceEvent type (D37630523 (https://github.com/facebookexperimental/eden/commit/dc3f77d92f4604b16ee28123d14b085ec343481e)). Building on top of these, this diff sets up the ActivityBuffer so that it subscribes to the InodeTraceBus to receive inode events instead of having events added directly.

Making this change is helpful as it will allow us to also easily generalize the ActivityBuffer in future diffs to contain fs and hg events as well which are already being published to a TraceBus.

In addition, by making this change the treeInode and fileInode unit tests need to get changed as Inode materialization events don't get added to the ActivityBuffer synchronously in the main Eden thread. Instead, in the unit tests we add a subscriber to the InodeTraceBus that reads all materialization events in the callback function and adds them into a synchronized queue that we read from to ensure materialization events are occurring in the order we expect.

Reviewed By: MichaelCuevas

Differential Revision: D37667501

fbshipit-source-id: c0fdbb57c02092b4899e1b2a88deb482b6eee05a