repo_bookmark_attrs: migrate BookmarkAttrs to a facet
Summary:
`BookmarkAttrs` are created at various times out of the repo bookmark config.  This is inefficient - we create new permissions checkers for *all* bookmarks even when landing to a single one.

Create a new repo attribute: `RepoBookmarkAttrs`.  This stores the bookmark attrs on the repo, allowing us to access it whenever it's needed.

Reviewed By: yancouto

Differential Revision: D37500880

fbshipit-source-id: ea6684820628bf6bf91bee2e0592f3d0ffffbb5d