testlib: add a method for listing files in the working copy
Summary: Getting a list of files seems in the current working directory seems to be a common operation so far in the new tests. Since we expect to do this more often using when using the new testing framework, adding a method for this in the working copy would help reduce boilerplate in the tests in the future.

Reviewed By: DurhamG

Differential Revision: D37512484

fbshipit-source-id: 975269e6c93479dbacf812d55bcdfc782cb77fb7