config: add Rust config command
Summary:
Currently our cloning scripts depend on calling `hg config` for determining whether the new codepath for invoking the Rust clone should be called or not. Running Python hg commands has an overhead of a few seconds, which is specially noticeable on Windows. Porting the config command to Rust should shave off a few seconds off from the Windows cloning scripts.

This diff adds the boilerplate required for the Rust clone command.

Reviewed By: DurhamG

Differential Revision: D37512486

fbshipit-source-id: a1484879bc3d6af2ed8115687846d6ac1ec1a199