smartlog: choose better GCA after max-commit-threshold limit
Summary:
The GCA choice uses a `public() &` optimization for performance. That
optimization depends on that the `parents(draft())` will include the
public nodes. However, after `max-commit-threshold` limit the
`parents(draft())` might be removed and therefore the optimization is
no longer sound.

Drop the optimization to make the logic more robust.

The optimization was important for revlog but less important for segmented
changelog. Practically, in my fbsource checkout with near 3k drafts, dropping
it only slows down the calculation from about 10ms to 13ms:

  In [1]: pg=repo.dageval(lambda: draft() | parents(draft()))
  In [2]: globals()['pg']=pg

  In [3]: %time repo.dageval(lambda: gcaall(public() & pg))
  Out[3]: CPU times: user 9.73 ms, sys: 0 ns, total: 9.73 ms
  Wall time: 9.68 ms
  <spans [6039ff447d472eb745ee2f8660cf460995538ecb]>

  In [6]: %time repo.dageval(lambda: gcaall(pg))
  Out[6]: CPU times: user 13 ms, sys: 0 ns, total: 13 ms
  Wall time: 13.1 ms
  <spans [6039ff447d472eb745ee2f8660cf460995538ecb]>

  In [8]: len(pg)
  Out[8]: 2833

Reviewed By: sggutier

Differential Revision: D37795117

fbshipit-source-id: 3821dddcd6f6b2e9fe3eec205f4aabddce2da15c