Fix up timestamps in remote gitimport
Summary:
I read the docs in source_control.thrift, and assumed that minutes was the right unit for offset. It's not.

Additionally, we appear to have misinterpreted git timestamps for some time, and not noticed because the timezone in test cases is UTC. Fix this up, too, and amend tests so that we see timezone issues in future.

Reviewed By: mitrandir77

Differential Revision: D37791764

fbshipit-source-id: abacde7206ce536bb3500edb5335734424c05eca