Add specific hook and conflict errors on repo_land_stack with reason only
Summary:
This adds specific exceptions to the SCS endpoint repo_land_stack on thrift, and makes sure to use those new exceptions correctly from code. It uses the refactors done in previous diffs to be able to use a custom error, not `ServiceError`, and convert it correctly to the thrift type.

For now, these error types only have a `reason` field with human readable exception description, but on the next diff I'll add more information to the exception so it can be parsed and properly used in code to pipe pushrebase exception through SCS.

Reviewed By: mitrandir77

Differential Revision: D37618460

fbshipit-source-id: 41ddaab9dea2374f4e65f7c25f8e9dd458d4d99e