Add LoggableError trait and use it
Summary:
This creates a `LoggableError` trait and starts using it on `log_result` function, which logs the SCS result to scuba, implementing it for ServiceError such that the behaviour doesn't change.

This will allow us to use errors other than `ServiceError` on our SCS endpoints, while still reusing the same common code for all the other endpoints.

Reviewed By: mitrandir77

Differential Revision: D37615973

fbshipit-source-id: c1b3839d7f7d3f384401aacbc716f74f88b3bb92