Add PushrebaseConflicts variant to MononokeError
Summary:
This adds a `PushrebaseConflicts` variant to `MononokeError`, makes sure we're populating it correctly when converting from `BookmarkMovementError`, and deal with it correctly on places that match on `MononokeError`.

For now it's not really used anywhere, but I'll use it in the next diffs to make SCS return a more specific error on land failures.

Reviewed By: farnz

Differential Revision: D37615453

fbshipit-source-id: ac6588bb38b78cb1934f49e4d4b804809551dd66