Use hidden error type on repo_land_stack
Summary:
This might seem like a minor change, but is only possible due to the refactors on D37615973.

It changes `repo_land_stack` from returning `Result<_, ServiceError>` to `Result<_, impl Into<service::RepoLandStackExn> + LoggableError>`, which in this case is an internal private type that is not exposed to the rest of the crate.

On this diff, this private type is super simple, just a wrapper around `ServiceError`, such that this diff works as a refactor, changing no behaviour. However, on the next diff I'll add more things to this wrapper, that will allow us to produce custom exceptions on the SCS response.

Reviewed By: RajivTS

Differential Revision: D37616435

fbshipit-source-id: 308c98b0ec9d30f2cab51eb8122ebe87df5b696a