testharness: fix TestMount::hasFileAt on Windows
Summary:
On Windows, FileInode::stat doesn't set the st_mode field as it's meaningless
for that platform, thus this function would always fail, even when the inode is
a file. A more straightforward way of writing this function is to simply test
the file type, which is available on all platforms.

Reviewed By: genevievehelsel

Differential Revision: D37804796

fbshipit-source-id: 586acebe231e7c431198f3a358fddd11e7f2e0bd