status: rust dispatch to eden, watchman or normal fs
Summary: Enable using all three of eden, watchman and physical (normal) fs for getting pending changes in status. For now we keep the actual 'dispatch' logic in Python. The caveat here is that the physical fs has a deadlock bug so we continue to default to the Python implementation for now, but we at least enable rust status on eden :).

Reviewed By: DurhamG

Differential Revision: D37820045

fbshipit-source-id: 3a2e85b93f21f7938bc1ff7a7807f5b34f3a6e2f