mononoke: blobrepo_hg: Make generic over facets
Summary:
`BlobRepoHg` provides a set of extension methods on top of `BlobRepo`
for hg specific functionality. The code is at a low layer in Mononoke and so
upper layers need to provide a `BlobRepo` in order to use the hg specific
functionality. This makes refactoring code to remove `BlobRepo` difficult.

This diff refactors `BlobRepoHg` so that it can be implemented for a type that
has the required facets, rather than just `BlobRepo` (although `BlobRepo` does
implement all of the necessary facets). In a later diff, I use this to remove
`BlobRepo` completely from the hg_sync job.

This diff implements this by bounding the implementation of the `BlobRepoHg`
trait to types that have the required facets for a specific method. For
example, a type that implements `ChangesetsRef` and `BonsaiHgMappingRef` can
call the `changeset_exists` method.

I added per-method trait bounds so that a type doesn't need to implement all
facets in order to call a single method.

This diff also removes the `get_filenodes` and `hg_mutation_store` methods as
the functionality they provide can be accessed through facets instead.

NOTE: In this diff I've also tried to remove all usages of `BlobRepo` from the
`blobrepo_hg` crate. However, removing them from `create_changeset.rs` was
pretty tricky and requires some more thinking. I'll come back to that later.

NOTE: I've not renamed `BlobRepoHg` in this diff in order to keep the changes
smaller, but am happy to do so in a later diff if people have better name
suggestions.

Reviewed By: markbt

Differential Revision: D37789413

fbshipit-source-id: c192fb871e3f5784930c02d9c341294340e5eda3