utils: refinement to ImmediateFuture::thenError
Summary:
While using thenError in the following diff, I realized that the thenError code
had a bug that failed compiling. The added test is an example of a similar code
that would fail to compile despite being perfectly valid.

As a bonus, this also makes thenValue slightly more efficient when dealing with
exceptions in the SemiFuture case.

Reviewed By: chadaustin

Differential Revision: D37824174

fbshipit-source-id: 2f759a65ebffce456f8f4f13ea4792ce078ae926