Use TraceBus for eden trace thrift
Summary:
The `eden trace thrift` command previously used some synchronous code in
ThriftLogHelper to maintain the set of outstanding requests.  This moves
accounting of outstanding requests into a TraceBus callback, which (1) allows
this work to be done on a background thread, and (2) will enable streaming of
ongoing Thrift requests in a following diff.

Reviewed By: chadaustin

Differential Revision: D37769506

fbshipit-source-id: 210b9db76ae29692ab376853c37b6d9b113bfdb4