inodes: no longer follow symlinks for gitignore
Summary:
Starting with Git 2.32
(https://github.com/git/git/blob/master/Documentation/RelNotes/2.32.0.txt#L7),
gitignore symlinks are no longer followed, and auditing the entire codebase
shows that this feature is not used (or easily replaced).

Removing support for symlinked gitignore makes us closer in behavior to Git,
but also allows the code to be simplified a bit. When diffing trees, inodes
were being loaded in order to read the gitignore files, which had a few
drawbacks:
 - Loading inodes causes slower updates on Windows and macOS,
 - The behavior was likely incorrect as the content of the gitignore read came
   from the working copy state, not from the trees diffed.
Since handling of symlinks was the major benefit of loading inodes and this
behavior is no longer needed, the store code can simply load the correct
gitignore from the ObjectStore, eliminating the drawbacks listed above.

Reviewed By: chadaustin

Differential Revision: D37355747

fbshipit-source-id: 569478a958ff01cf9fbef01508008d8b29e0c056