Skip checkout action if inode matches destination already
Summary:
During a checkout, if the unmaterialized inode hash already matches the
destination hash of the checkout, then we don't need to take any action. This is
needed for the next diff where we allow interrupted checkouts to be resumed. In
that situation some of the inodes may already match the destination.

Reviewed By: chadaustin

Differential Revision: D37188338

fbshipit-source-id: 16c58101f5ce75d22d9d1766686417c990f00216