Expose TreeLookupCallback to OverlayChecker
Summary:
In a future diff we'll want the OverlayChecker to fix up the overlay by
dematerializing missing nodes, instead of creating empty files. To do that,
we'll need the ability to lookup the scm hash of a directory entry. This first
diff threads a lookup function from EdenMount down to the OverlayChecker. A
later diff will then use this callback.

Reviewed By: xavierd

Differential Revision: D37188336

fbshipit-source-id: 003143391de0fcf1b02ecc0c1336c0b7e88ec5d9