config: fix use of uninitialized member in testharness
Summary:
The MountProtocol in CheckoutConfig may be read on macOS to test if an NFS
mount should be created instead of a FUSE mount. Unfortunately, since
testharness doesn't create valid config, the
CheckoutConfig::loadFromClientDirectory function is not used, and thus the
mountProtocol_ field is never initialized.

In practice, the test InodeMap.invalidInodeNumber was flaky on macOS due to
this issue.

Reviewed By: kmancini

Differential Revision: D37841713

fbshipit-source-id: 49b160869d0e682b09ebd246b94e6f4dc999a8e5