Spawn bonsai_diff_unfold tasks in bonsai_diff
Summary:
This might help with the deadlock problems, and Simon says that in his testing, it made the import a lot faster anyway.

That being said, I think it might be better to do the spawning inside bounded_traversal_stream instead, in order to reduce the risk of deadlocks no matter how they are used.

Reviewed By: farnz

Differential Revision: D37839553

fbshipit-source-id: 6c1db9e5c4442b2526ec4d1cf66afc7ca28e391e