types: include more detail in network errors
Summary:
Include the wrapped error's Debug info in the outer error message.

I originally didn't do this because I wanted to rely on the error chain to show the inner error, but that just isn't reliable since it depends on how the error is printed. Now we always include the inner error's info which means we will see duplicative output when formatting with Debug.

Reviewed By: DurhamG

Differential Revision: D37854809

fbshipit-source-id: febf371daac26542507de86a01d3bfc434cc627e