fix mutable renames log
Summary:
The current version of mutable renames log code works only if the mutable
rename is directly attached to any commit from the immutable history. This
commits attempts to change that.

This diff changes our approach from fetching mutable renames as we traverse the
history graph to doing it in batches when we're processing fastlog batches.

Reviewed By: farnz

Differential Revision: D37782731

fbshipit-source-id: d42b94a4eaf7f11694203f54898c5ee13734558f