config: implement basic config display in the Rust config command
Summary:
The config command has two main features: displaying the value of configs and launching an editor for editing configs. This commit implements the former, but support for templates and the `--debug` option (which shows the origin of each config) is still pending.

In the new implementation sections and individual items are sorted alphabetically instead of using the default ordering the Python version had.

Reviewed By: DurhamG

Differential Revision: D37512487

fbshipit-source-id: 0d6d0a6f017c46544afcdb19a28dd9fd2cf7fd35