Enable activityBufferConfig by default
Summary: With the ActivityBufferBenchmark set up showcasing reasonable performance overheads for writes with the ActivityBufferBenchmark enabled, it is time to enable the ActivityBuffer for all production eden instances. However, we still keep the config so that we have a knob to turn off in case anything goes wrong. After this, we will soon remove the chef recipe that enables the ActivityBuffer to true specifically for the Eden team, and then afterwards remove the config altogether.

Reviewed By: chadaustin, genevievehelsel

Differential Revision: D37887944

fbshipit-source-id: de70a69eaca1c0efe3a5ffa0fca7bc9f56c587ab