Create Performance Benchmarks to measure ActivityBuffer overhead
Summary:
In order to help with retroactive debugging of EdenFS, we recently set up support for an eden trace inode --retroactive command to retrieve information about the last n inode materializations. In order to do this, an ActivityBuffer in each EdenMount subscribes to an InodeTraceBus to receive and store information about inode materializations as they occur to be viewed retroactively by the eden trace inode --retroactive command.

In this diff, we set up performance benchmarks that measure the time needed to publish Inode events to an EdenMount's inodeTraceBus as well as read these events into the subscribed ActivityBuffer. This is done through the materialization paths of creating a directory and writing to a file where we expect activitybuffer overheads to be relatively largest. Specifically, the benchmark first times creating a large number of files, then unmaterializes the files by committing them, checking out the base parent commit, and checking back into the new commit, and then times writing to each file. Additionally, the unmaterializing and writing portion of the benchmark is repeated through several iterations. This benchmark is run both with the ActivityBuffer enabled and disabled to measure the overheads from adding to the ActivityBuffer.

Also, this benchmark currently does not run on windows due to folly include errors, though a task is set up to fix this in the future.

Reviewed By: chadaustin

Differential Revision: D37857084

fbshipit-source-id: 0bbde7d66ff07cb22546bb5fdbaadc4c1494b60c