Dynamically adjust inode width in trace inode command
Summary: Before this diff, a UI bug was present in the `eden trace inode` and `eden trace inode --retroactive` command where when inodes would be 6 digits or longer they would occupy more space than designated for the inode column of the ouput as per the image below. This diff fixes this issue by keeping track of the most amount of digits either provided in the retroactive command or seen up to now in the live version and uses that as a fixed quantity for the inode column width (up to a minimum width of 5 digits still). An example after this diff is also shown below{F752750723} {F752758358}

Reviewed By: genevievehelsel

Differential Revision: D37895413

fbshipit-source-id: d537704e07e0e6fb0bea9e3b0adaea5312f5bd7b