only fetch nessecary data for entry attributes
Summary:
readdir and getEntryAttributesForPath allow collecting all the attributes for a
directory or a list of files, but this currently always fetches all the
attributes no matter which are requested.

Buck2 will only need a subset of the attributes on the readdir call.

So its worth optimizing here to only fetch requested attributes.

There are still some optimizations work making:
- we should avoid looking up InodeOrTreeOrEntry when no attributes are requested
- we can even probably avoid looking up some data when requesting the
TreeEntryType only.

Reviewed By: xavierd

Differential Revision: D37557706

fbshipit-source-id: 894dd9907e55f7255473fd9564812e2bb773c45d