Mononoke Sharded Process Manager: Healing Unhealthy Repos
Summary:
The current sharded model caters for complete failure scenario, i.e. if the underlying host serving the shards/repos fails then all of the repos are rescheduled for execution and TW takes care of restarting the host. However, since the sharded model allows for co-location of multiple repos on the same server it is possible to have transient partial failures, i.e. one of the repos errors-out while the remaining keep executing without failures.

This diff includes changes to help detect such scenarios and to provide the right signal to SM so that it can drop and re-add this repo essentially achieving the affect of a task-restart.

Differential Revision: D37901000

fbshipit-source-id: cb20f9588b460c9b22a143d2a9216c7ad9f0daa4