thrift readdir fix segfault
Summary:
edenfs can crash when we attempt to load inodes for TreeInode::getChildren.

This is because the wlock side is implicitly unlocking the lock, so subsequent
lookups do not have valid access to the tree inode's contents.

to remove this unlock, we need to move the code that ran after the unlock to
getChildren, because this code needs to execute outside of the lock.

Reviewed By: chadaustin

Differential Revision: D37835806

fbshipit-source-id: a2b09aed9003a57cb7ab5c530bb4d9b99eef3bcb