Fix wording in Inode Number Allocation eden documentation
Summary: The EdenFS Inodes documentation currently incorrectly suggests that listing a directory materializes it. While it does write the directory to the overlay in order to record the inode numbers assigned to its children, it doesn't actually materialize the directory since a source control object ID is still preserved and used in order to get the directory's contents. This change clarifies this wording by only mentioning how the directory is written to the overlay and not materialized to it.

Reviewed By: xavierd

Differential Revision: D37975525

fbshipit-source-id: 22a6119940206d2a26243d6f3ccc7b883fe05fa7