Add monitoring_bookmark to metaconfig
Summary:
Problem: Currently, when we log pushrebase metrics (for calculating saturation), we do so on all pushrebase operations, independent of in which bookmark they're pushrebasing into.

This is overcounting the saturation, as two pushrebase opeations on different bookmarks can happen in parallel.

This stack aims to fix that, by only logging for a specific bookmark (We could also log for all bookmarks, but the data for most bookmarks is not that useful, as usually only one of them happens a lot).

This diff only reads the field added to configerator and parses it to the correct struct, but it is still unused and will be used on the next diff.

Reviewed By: HarveyHunt

Differential Revision: D37787218

fbshipit-source-id: 0e4ab7b5f8828f94e4343446fa195d9e0b2cb37f