Don't reimport known git commits
Summary:
Right now, if `git` tells us that we need to import a commit, we do so blindly.

Change this so that we check to see if Mononoke knows about a commit before importing it - that way, we don't bother with file blob uploads etc for commits we know about.

The `gitimport` command has an override to forcibly import a commit if we need it.

Reviewed By: mzr

Differential Revision: D37933345

fbshipit-source-id: e51a818fb0986a285c9e03029f13f387e719002b