Add hasChild to TreeOverlayStore
Summary:
A later diff will want to update overlay entries in the tree overlay. To do so, we need to first check if the entry exists and remove it if so. This diff adds the hasChild check which will let us decide whether to remove the existing entry.

Note, there was a hasChild query already, but it actually checked if the given (parent, childname) child directory has children, vs if the parent had the specified child. Given that addChild and removeChild already exist and operate on the passed in parent/child relationship, it makes sense for hasChild to test for that same relationship. I moved the older hasChild logic to be under hasChildren.

Reviewed By: fanzeyi

Differential Revision: D37694131

fbshipit-source-id: bf359534bee3ce5e24c8c2e8e939dba866ad5fdf