inodes: print conflicts during checkout
Summary:
Based on manual observation, it looks like Mercurial completely ignores
conflicts/errors during force/normal checkout operations. I have a suspicion
that this could explain some users having a large amount of changes in
`hg status` after update/rebase on Windows. To validate this, let's both log
the number of conflicts in the edenfs.log and in the telemetry.

Reviewed By: kmancini

Differential Revision: D37940852

fbshipit-source-id: 71caf0b899d76a44bfce1d02d279238826d972e7