tweakdefaults: allow using the original pull command to handle --update
Summary:
The original pull command can handle `pull -B x --update` or `pull -r y
--update` fine. However, with tweakdefaults, it will force the `--update`
destination to the `tweakdefaults.defaultdest` config and there is no way to
use the more desired original pull logic.

Change it so if `tweakdefaults.defaultdest` is empty, use the original pull
command to handle `--update`. This is especially useful when
`tweakdefaults.defaultdest` does not exist in the repo (ex. `remote/master`
does not exist in git repos using the `main` as the main branch name).

Differential Revision: D38066104

fbshipit-source-id: 26700b8c1859ba7a46fd43e197ac18fb091cc42c