inline in_throttled_slice in create_warm_bookmark_cache_scuba_sample_builder
Summary: IMO it's better this way as `in_throttled_slice` looked like it was more about dynamically throttling single requests (that's what is means in the rate_limitting crate). This is just a 3 line function used in only one place (although 2 times due to code duplication because of the ongoing migration to new cmdlib), and I think it looks better this way.

Reviewed By: HarveyHunt

Differential Revision: D38112458

fbshipit-source-id: 71b46eeb37b347d89e4bf16aec62c2ebb9a96ade