Use filenames in inode tracing commands when full path isn't available
Summary: In a previous change (D37961108 (https://github.com/facebookexperimental/eden/commit/18bc0ee56a5c7f8939e8e08b104bf74d44dd1da5)), we added path information to the `eden trace inode` live and retroactive commands. In that change, the path information was computed only when needed in the thrift Eden service handler using the current EdenMount's InodeMap. However, when doing this, retroactive path information may not always be available (or correctly representative of the past path) if an inode in the path has since been removed (or changed), and live path information may not always be available in the InodeMap in the case of start events for Inode creations. This diff addresses this issue by adding in the current inode's filename (but not the full path) to each tracebus InodeEvent when the InodeEvent occurs and defaults to displaying just this filename if the full path info isn't available. An exception occurs when new inodes are created and we compute the full path anyway in which case we do add the full path to the tracebus InodeEvent.

Reviewed By: kmancini

Differential Revision: D38000772

fbshipit-source-id: 7a718989c397425cc0b0c96def21fd09b5b55fc9