Add full path info in inode trace bus subscription
Summary:
This change finalizes our design for adding path information to the EdenFS inode tracing CLI. After experimenting with designs that
* First: calculated path information only in the eden thrift service handler when CLI calls were made (this is best performance-wise but could lead to potential accuracy problems)
* Second: calculated path information inline during inode materializations (best accuracy but can potentially harm performance with large file trees)

We are deciding to go with a third middle ground approach that calculates path information upon tracebus subscriptions. In this way calculating paths is offloaded from the critical materialization path but is still done soon after materialization occurs so accuracy holds the vast majority of the time. Note, for the live inode tracing command this means we continue to calculate paths in the service handler as the handler subscribes to the inode tracebus.

In addition, we continue to calculate file names inline during materializations (and in fact full paths during inode creations when they are readily available), which we use as a default value even if our later path calculation fails.

Reviewed By: kmancini

Differential Revision: D38083187

fbshipit-source-id: 88702f98d64998be1a87c031de67186f3e233366