configparser: move parsing logic to a separate crate
Summary:
The configparser crate is bloated. Move the pest related parser to a separate
crate so the parsing logic can be reused.

The standalone parsing logic can be potentially useful in just looking at a
repo's hgrc's paths.default config to figure out the repo name, without a feature-complete ConfigSet.

Reviewed By: DurhamG

Differential Revision: D37702817

fbshipit-source-id: a26d8b439158637f3cbd26b59f5718cdc1ddc876