capability matching for takeover
Summary:
We are migrating to capabilities because making changes with version numbers
means adding a lot of opaque version checks in the code. Capabilities are
a bit more self documenting.

Let's move one more step towards capabilities by introducing capability
matching between the client and server.

The next step will be to remove all the version number code. But this diff needs
to be stable before we can do that.

Reviewed By: chadaustin

Differential Revision: D32226963

fbshipit-source-id: 951da5d35cfb130535b7052869ac9722854198cc