eagerepo: support eager repo as a local repo
Summary:
Previously eager repo can only be used as a remote peer. This diff makes it so
eager repo can also be used as a local repo. Under the hood, the dag/hgcommits,
metalog use the same on-disk format as regular client repos. The "store" uses
a same format as the hgcommits' SHA1 `key -> [p1][p2][content]` format, backed
by the EagerRepoStore object.

Supporting it is similar to supporting the git store.

Reviewed By: DurhamG

Differential Revision: D38142905

fbshipit-source-id: 0c60ea16f7ce7831061837825a4495b81275c519