paths: in debug/sanitized builds pessimize moves to copies
Summary:
In D37422971 (https://github.com/facebookexperimental/eden/commit/3d9636fe8c7505642e3e852218827f0250b99ff9), it was shown that moving a path will invalidate the pieces on the
moved path. Unfortunately, this can only be observed when the path is short
enough to fit in the SSO of std::string. To force this to be true even without
SSO, we can pessimize moves into copies and leverage the use of ASAN to tell us
if the code is misbehaving.

Reviewed By: kmancini

Differential Revision: D37705144

fbshipit-source-id: a4446632b0e5d3b422226e85c324059c5b8cf379