Return RelativePath() in InodeMap if getting path for the root
Summary:
When an eden mount is in the process of unmounting, the inodeTraceBus for that mount may continue to publish events to the ActivityBuffer. When doing so, it will get paths for the relevant inodes using the mount's InodeMap. However, there is a bug here where the unmounting causes the root inode to be freed, yet it remains in the loadedInodes_ struct of the InodeMap. When the path for this root inode is called for by the tracebus subscription, a use after free will run as the path finding code will try to access information about the freed root inode that the InodeMap still has in its loadedInodes_ struct.

As a temporary fix, we return RelativePath() at the top of the InodeMap's path finding code if the path for the root inode is requested until we further investigate why the root inode is still in the loadedInodes_ struct of the inodeMap.

Reviewed By: xavierd

Differential Revision: D38218783

fbshipit-source-id: 062c47f9eab192f2f60e6de45e563c9027eca2bd