inodes: use ManualExecutor in inode tests
Summary:
In a future subsequent change, ImmediateFuture will be taught to always be lazy
in debug/sanitized builds in order to reduce the chance of use after free bugs
slipping through tests. With that diff, ImmediateFuture::isReady will always
return false, and is thus not a good way to test that data is immediately
available. A better way is to extract a SemiFuture and attach it to a
ManualExecutor. That way the test can drain it when needed which could avoid
races in the case where some underlying continuation hasn't executed due to
timing issues.

Reviewed By: kmancini

Differential Revision: D37729624

fbshipit-source-id: 07f7ec75d30493cb3f1ffaf840bb6dc88b253fe1