Delete old command and replace it with new cmdlib command
Summary:
The previous diff finished moving scsc commands to use the new command lib! Now we need to swap the two implementation behind the scenes, so that everything that is using the old SCSC target now points to the new version.

This is done in two steps:
1. This diff moves the scsc_new files under a eden/mononoke/scs/client/src_v2 directory, deletes the old code, and fixes all temporary code to stop using "scsc_new" instead of "scsc".
2. The next diff moves `src_v2` to `src`, so things go back to where they used to live.

This needs to be done in two diffs because, as far as I can tell, it's not possible to delete a file and at the same time replace it with the move of another file. Also, we need to move things to `src_v2` inside the same directory as `src` because buck can't handle rules that depend on stuff outside the `TARGETS` file directory.

Here we want `client_new/src/main.rs` --> `client/src/main.rs`, but can't to that directly so instead do `client_new/src/main.rs` --> `client/src_v2/main.rs` --> `client/src/main.rs`.

Reviewed By: mitrandir77

Differential Revision: D38029238

fbshipit-source-id: 3d7e8003b460670bc18e98baae8c8c0c5d6a40fe