mononoke: lfs: Remove BlobRepo from lfs_server
Summary:
The LFS server needs access to a filestore config and a blobstore,
however it has been using a full `BlobRepo`. Now that we have facets, the LFS
server can be refactored to require just a few facets.

Update the code to remove dependencies on `BlobRepo`.

NOTE: In a follow up diff I'll add a permission checker into the `Repo` struct
rather than defining a custom one that is stored alongside `Repo`.

Reviewed By: markbt

Differential Revision: D38209865

fbshipit-source-id: d138589175e08aae4bcedab4d4a29761154d6f82