Remove root from inodeMap loadedInodes struct when freed
Summary:
We recently noticed when getting the path for the root inode in the ActivityBuffer tracebus subscription function, that the root inode would remain in an InodeMap's loadedInodes_ struct even after the root inode gets freed when the InodeMap is shutting down. This can cause use after free errors if not dealt with carefully.

To prevent this, we remove the root Inode from the loadedInodes_ struct immediately before freeing it

Reviewed By: chadaustin

Differential Revision: D38227083

fbshipit-source-id: e7f5d8968f29dc9816e3777b432bd0d6f4bd9142