Track loads in eden inode tracing cli
Summary:
So far the ActivityBuffer in EdenFS has only supported storing inode materialization events, which are the only type of event currently being displayed in the eden trace inode and eden trace inode --retroactive commands. In this change and the previous diff, we add support for storing and displaying inode load events as well.

This change specifically modifies the InodeMap code to publish events when inodes first begin loading, finish loading, or fail loading, as well as adds a new attribute on UnloadedInode objects to record the timestamp for when the inode first started loading.

Reviewed By: kmancini

Differential Revision: D38092233

fbshipit-source-id: 4a16e811d596db0553b3ca4d503e4843d82fc6c4