treemanifest: avoid unsound `rev(tip) - int` calculation
Summary:
The calculation is not sound. For example, if the `int` is too large, the
revision would be negative and error out. With segmented changelog,
if `tip` is a non-master commit, and `int` exceeds the number of non-master
commits, it will produce an invalid revision number and error out.

Change it to use the `limit` revset instead of operating on rev numbers
directly to be more robust.

Differential Revision: D38233264

fbshipit-source-id: 63e07b8fd945f099ebd8031c19e4f1e61467d780