Use MononokeApp for all source-control services initialization
Summary:
This diff is part of changes for project `Convergence`, where `MononokeApp` will end up becoming the single point of environment and state access for all services and commands in source control. This diff includes the following changes:
- Removed `MononokeApiEnvironment` and transferred the arguments contained in it to `MononokeEnvironment` which is part of `MononokeApp`
- Added `filter-repos` arg as an `AppExtension` in `MononokeApp` instead of keeping it as a direct clap argument in relevant commands / services
- Removed unnecessary dependencies generated as part of this change.

Unfortunately splitting this diff was a real challenge and it lead to circular dependencies so kept it all together.

Reviewed By: HarveyHunt

Differential Revision: D38254837

fbshipit-source-id: ef7ada6b376f2da1ebdcf901e21322698ab6083b