configparser: handle config overrides in hg::load()
Summary:
Move the "--config" and "--configfile" logic directly into hg::load() so callers don't have to worry about the subtle ordering issues. I tried to use the "readonly" (AKA "pinned") facility, but that does not interact correctly with the superset verification stuff. Instead, we just load the config overrides before and after loading the rest of the config.

We currently have a bug in Rust where "--config" values are not available in the config as the config is being loaded (because we insert them after the config has finished loading). In particular, some dynamic config loading logic can be configured via the config.

I will make use of the new functionality in a following diff.

Reviewed By: quark-zju

Differential Revision: D38127037

fbshipit-source-id: c3b2d181031e4489a5e27a0ac7237a986ed19b93