commitcloud: track all workspace sync states in a single file
Summary:
Previously, commitcloud state uses one file per workspace. That's not friendly
for metalog use-case where the filenames are better fixed. Change SyncState
to use a single file for all workspace states to address the problem.

For now we continue writing the old state files for compatibility. But new
state files are used as the "source of truth" once the old file gets migrated
to the new format.

In the future we might want to delete writing support for the old state files,
then delete reading support.

test-commitcloud-sync.t is slightly updated so it waits for the background
"rebase" command to complete. Otherwise the test might fail with
"repo changed while backing up".

Reviewed By: markbt

Differential Revision: D35232804

fbshipit-source-id: b56b82417d7b13ed4cedbf1a297ca94b8238f9a6